Change the type amount field from int to float

+++ src/app/services/transaction.ts
@@ -71,7 +71,7 @@ const extendedApi = api.injectEndpoints({
         query: (args) => ({
           document: gql`
             mutation AddTransaction(
-              $amount: Int!
+              $amount: Float!
               $date: String!
               $userId: String!
               $categoryId: String!
